Move demand slicing code into own submodule
